refactor(store): move types to their file and improve typings in @Debounce decorator
